Committing first experimentation with separating the detector array for each work-item.  That is, each work-item indexes into the global_results array to access it's individual portion of the array. Ensures no issues with atomic updating of values.
Strange behavior still persists however.  Launching a single work-item
to simulate many photons produces different results from launching
many work-items to simulate a single photon.  It seems that if both
simulate the same number of photons the results should have a high
degree of correlation, but that is not the case.

To do:
Give each work-item a different mu_s and verify that each produces
results that differ from one another.  Currently, when launching many
work-items to simulate a single photon they all produce the same
result.  This removes the advantage of highly-parallel hardware and is
obviously undesired.